Updated YuGiOh card upload script to use general script, and renamed fil…
