+++ DownloadImages/UploadFromApiToS3.py
@@ -10,7 +10,7 @@ endpoint: (String): The api endpoint being used to gather images
 query_params: (Dict): Paramaters used to filter the api response
 response_to_image_urls: (Function): Recieves the api response, and converts it into a list of image urls
 '''
-def download_images_s3(bucket_name, endpoint, query_params, json_to_image_urls):
+def upload_api_images_to_s3(bucket_name, endpoint, query_params, json_to_image_urls):
     s3 = boto3.client('s3')
     create_bucket_if_not_exists(s3, bucket_name)
     resp = requests.get(endpoint, query_params)
@@ -25,15 +25,19 @@ def download_images_s3(bucket_name, endpoint, query_params, json_to_image_urls):
         os.makedirs(dir)
 
     for image_url in image_urls:
-        filename = dir + '/' + image_url.split("/")[-1]
+        filename = image_url.split("/")[-1]
+        full_path = dir + '/' + filename
 
         r = requests.get(image_url, stream = True)
 
         if r.status_code == 200:
             r.raw.decode_content = True
 
-            with open(filename, 'wb') as f:
+            with open(full_path, 'wb') as f:
                 shutil.copyfileobj(r.raw, f)
+                s3.upload_file(full_path, bucket_name, filename)
+        else:
+            print("Image url {} failed with response code {}!".format(image_url, r.status_code))
 
     shutil.rmtree(dir)
 
@@ -48,15 +52,3 @@ def create_bucket_if_not_exists(client, bucket_name):
             break
     if not bucket_exists:
         client.create_bucket(Bucket=bucket_name)
-
-blue_eyes_params = {
-    'archetype': 'Blue-Eyes'
-}
-
-def yugioh_func (json):
-    url_list = []
-    for card_data in json.get('data'):
-        url_list.append(card_data.get('card_images')[0].get("image_url"))
-    return url_list
-
-download_images_s3("blue-eyes2", "https://db.ygoprodeck.com/api/v7/cardinfo.php", blue_eyes_params, yugioh_func)
